pyproject.toml
[project]
name = "test-bank-webhook"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "django>=5.2.1",
    "django-environ>=0.12.0",
    "djangorestframework>=3.16.0",
    "drf-spectacular>=0.28.0",
    "psycopg2>=2.9.10",
]

[dependency-groups]
dev = [
    "coverage>=7.8.2",
    "pytest>=8.3.5",
    "pytest-cov>=6.1.1",
    "pytest-django>=4.11.1",
    "pytest-mock>=3.14.1",
]
